separate commands and capability
