+++ lib/appium_lib.rb
@@ -7,5 +7,5 @@ require 'forwardable' unless defined? Forwardable
 # Init global driver
 $driver = nil
 
-require_relative 'appium_lib/logger'
+require_relative 'appium_lib/core/logger'
 require_relative 'appium_lib/appium'
