+++ lib/appium_lib/core/core.rb
@@ -5,6 +5,7 @@ require_relative 'patch'
 require_relative 'search_context'
 require_relative 'wait'
 require_relative 'device'
+require_relative 'capabilities'
 
 require_relative 'common/log'
