+++ lib/appium_lib/core/logger.rb
