+++ lib/appium_lib/ios/command.rb
@@ -3,7 +3,6 @@ module Appium
     module Commands
       COMMAND_NO_ARG = {
         device_time:                [:get,  'session/:session_id/appium/device/system_time'.freeze],
-        current_context:            [:get,  'session/:session_id/context'.freeze]
       }.freeze
 
       COMMAND = {
